split up made doorway work
