+++ apocalypse_sim/agents/agent.py
@@ -1,6 +1,6 @@
 from mesa import Model
 from mesa import Agent as MesaAgent
-from .map_object import Road
+from grid_map.map_object import Road
 from shapely.geometry import Point
 
 class Agent(MesaAgent):
