+++ apocalypse_sim/automaton/automaton.py
