+++ apocalypse_sim/automaton/state.py
