+++ apocalypse_sim/grid_map/data/scrapy.py
