+++ apocalypse_sim/grid_map/map_object.py
