--- apocalypse_sim/netherlands_data/provinces_densities.csv
@@ -1,13 +1,0 @@
-,statnaam,population,surface,population_per_km,density_of_total
-0,Groningen,582649.0,2333.28,250,0.040703353956366005
-1,Friesland,646092.0,3341.7,193,0.031422989254314554
-2,Drenthe,488871.0,2641.09,185,0.030120481927710843
-3,Overijssel,1142360.0,3325.62,344,0.05600781504395962
-4,Flevoland,403280.0,1417.5,285,0.04640182351025725
-5,Gelderland,2031123.0,4971.76,409,0.06659068707261478
-6,Utrecht,1268489.0,1385.02,916,0.14913708889612504
-7,Noord-Holland,2775617.0,2671.03,1039,0.16916313904265712
-8,Zuid-Holland,3607150.0,2814.69,1282,0.20872679908824487
-9,Zeeland,381182.0,1787.13,213,0.03467925757082384
-10,Noord-Brabant,2495107.0,4916.49,507,0.08254640182351025
-11,Limburg,1115805.0,2150.87,519,0.08450016281341582
